Wrote Java Intergration
This object reads values from a file. This file will be generated using
a COM port printing program, this is to allow for the software to work a
multiple platforms because I could not find a mac COM port library for
JAVA. This object requires that the file path be passed to it as a
string when being constructed. The object will be interacted with by
calling the readSensor() function and this function will return the int
that is listed in the file. I have not been able to fully test this
object.